--- libspu/mpc/swift/protocol_single_test.h
@@ -1,14 +1,0 @@
-#include "gtest/gtest.h"
-#include "yacl/link/link.h"
-
-#include "libspu/mpc/api_test_params.h"
-
-namespace spu::mpc::test {
-
-class ArithmeticTest : public ::testing::TestWithParam<OpTestParams> {};
-
-class BooleanTest : public ::testing::TestWithParam<OpTestParams> {};
-
-class ConversionTest : public ::testing::TestWithParam<OpTestParams> {};
-
-}  // namespace spu::mpc::test
