+++ src/libspu/mpc/swift/boolean.h
@@ -1,4 +1,4 @@
-// Copyright 2021 Ant Group Co., Ltd.
+// Copyright 2024 Ant Group Co., Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,6 +17,29 @@
 #include "libspu/mpc/kernel.h"
 
 namespace spu::mpc::swift {
+
+class CommonTypeB : public Kernel {
+ public:
+  static constexpr const char* kBindName() { return "common_type_b"; }
+
+  ce::CExpr latency() const override { return ce::Const(0); }
+
+  ce::CExpr comm() const override { return ce::Const(0); }
+
+  void evaluate(KernelEvalContext* ctx) const override;
+};
+
+class CastTypeB : public CastTypeKernel {
+ public:
+  static constexpr const char* kBindName() { return "cast_type_b"; }
+
+  ce::CExpr latency() const override { return ce::Const(0); }
+
+  ce::CExpr comm() const override { return ce::Const(0); }
+
+  NdArrayRef proc(KernelEvalContext* ctx, const NdArrayRef& in,
+                  const Type& to_type) const override;
+};
 
 class B2P : public UnaryKernel {
  public:
@@ -83,6 +106,8 @@ class AndBB : public BinaryKernel {
   ce::CExpr latency() const override { return ce::Const(0); }
 
   ce::CExpr comm() const override { return ce::Const(0); }
+
+  Kind kind() const override { return Kind::Dynamic; }
 
   NdArrayRef proc(KernelEvalContext* ctx, const NdArrayRef& lhs,
                   const NdArrayRef& rhs) const override;
