+++ src/libspu/mpc/swift/io_test.cc
@@ -1,4 +1,4 @@
-// Copyright 2021 Ant Group Co., Ltd.
+// Copyright 2024 Ant Group Co., Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,21 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "libspu/mpc/swift/type.h"
+#include "libspu/mpc/io_test.h"
 
-#include <mutex>
-
-#include "libspu/mpc/common/pv2k.h"
+#include "libspu/mpc/swift/io.h"
 
 namespace spu::mpc::swift {
 
-void registerTypes() {
-  regPV2kTypes();
-
-  static std::once_flag flag;
-  std::call_once(flag, []() {
-    TypeContext::getTypeContext()->addTypes<AShrTy, BShrTy, PShrTy>();
-  });
-}
+INSTANTIATE_TEST_SUITE_P(
+    SwiftIoTest, IoTest,
+    testing::Combine(testing::Values(makeSwiftIo),  //
+                     testing::Values(3),            //
+                     testing::Values(FieldType::FM32, FieldType::FM64,
+                                     FieldType::FM128)),
+    [](const testing::TestParamInfo<IoTest::ParamType>& p) {
+      return fmt::format("{}x{}", std::get<1>(p.param), std::get<2>(p.param));
+    });
 
 }  // namespace spu::mpc::swift
