+++ src/libspu/mpc/swift/protocol.h
@@ -1,0 +1,30 @@
+// Copyright 2024 Ant Group Co., Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "yacl/link/context.h"
+
+#include "libspu/core/context.h"
+
+namespace spu::mpc {
+
+void regSwiftProtocol(SPUContext* ctx,
+                      const std::shared_ptr<yacl::link::Context>& lctx);
+
+std::unique_ptr<SPUContext> makeSwiftProtocol(
+    const RuntimeConfig& conf,
+    const std::shared_ptr<yacl::link::Context>& lctx);
+
+}  // namespace spu::mpc
